refactor: api mangers 와 guests로 분리
