+++ backend/src/main/java/com/woowacourse/zzimkkong/domain/Manager.java
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Provider {
+public @interface Manager {
 }
