+++ backend/src/main/java/com/woowacourse/zzimkkong/service/ManagerReservationService.java
@@ -1,14 +1,16 @@
 package com.woowacourse.zzimkkong.service;
 
+import com.woowacourse.zzimkkong.domain.Map;
+import com.woowacourse.zzimkkong.domain.Member;
 import com.woowacourse.zzimkkong.domain.Reservation;
 import com.woowacourse.zzimkkong.domain.Space;
 import com.woowacourse.zzimkkong.dto.reservation.ReservationCreateResponse;
+import com.woowacourse.zzimkkong.dto.reservation.ReservationCreateUpdateRequest;
 import com.woowacourse.zzimkkong.dto.reservation.ReservationCreateUpdateWithPasswordRequest;
-import com.woowacourse.zzimkkong.dto.reservation.ReservationPasswordAuthenticationRequest;
 import com.woowacourse.zzimkkong.dto.reservation.ReservationResponse;
 import com.woowacourse.zzimkkong.dto.slack.SlackResponse;
+import com.woowacourse.zzimkkong.exception.authorization.NoAuthorityOnMapException;
 import com.woowacourse.zzimkkong.exception.reservation.NoSuchReservationException;
-import com.woowacourse.zzimkkong.exception.reservation.ReservationPasswordException;
 import com.woowacourse.zzimkkong.exception.space.NoSuchSpaceException;
 import com.woowacourse.zzimkkong.repository.MapRepository;
 import com.woowacourse.zzimkkong.repository.ReservationRepository;
@@ -16,20 +18,21 @@ import com.woowacourse.zzimkkong.repository.SpaceRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Service
 @Transactional
-public class UserReservationService extends ReservationService {
-    public UserReservationService(
-            final MapRepository maps,
-            final SpaceRepository spaces,
-            final ReservationRepository reservations) {
+public class ManagerReservationService extends ReservationService {
+    public ManagerReservationService(MapRepository maps, SpaceRepository spaces, ReservationRepository reservations) {
         super(maps, spaces, reservations);
     }
 
     public ReservationCreateResponse saveReservation(
             final Long mapId,
-            final ReservationCreateUpdateWithPasswordRequest reservationCreateUpdateWithPasswordRequest) {
+            final ReservationCreateUpdateWithPasswordRequest reservationCreateUpdateWithPasswordRequest,
+            final Member manager) {
         validateMapExistence(mapId);
+        validateAuthorityOnMap(mapId, manager);
 
         validateTime(reservationCreateUpdateWithPasswordRequest);
         Space space = spaces.findById(reservationCreateUpdateWithPasswordRequest.getSpaceId())
@@ -53,34 +56,35 @@ public class UserReservationService extends ReservationService {
     public ReservationResponse findReservation(
             final Long mapId,
             final Long reservationId,
-            final ReservationPasswordAuthenticationRequest reservationPasswordAuthenticationRequest) {
+            final Member provider) {
         validateMapExistence(mapId);
+        validateAuthorityOnMap(mapId, provider);
 
         Reservation reservation = reservations
                 .findById(reservationId)
                 .orElseThrow(NoSuchReservationException::new);
-        checkCorrectPassword(reservation, reservationPasswordAuthenticationRequest.getPassword());
         return ReservationResponse.from(reservation);
     }
 
     public SlackResponse updateReservation(
             final Long mapId,
             final Long reservationId,
-            final ReservationCreateUpdateWithPasswordRequest reservationCreateUpdateWithPasswordRequest) {
+            final ReservationCreateUpdateRequest reservationCreateUpdateRequest,
+            final Member provider) {
         validateMapExistence(mapId);
-        validateTime(reservationCreateUpdateWithPasswordRequest);
+        validateAuthorityOnMap(mapId, provider);
+        validateTime(reservationCreateUpdateRequest);
 
-        Space space = spaces.findById(reservationCreateUpdateWithPasswordRequest.getSpaceId())
+        Space space = spaces.findById(reservationCreateUpdateRequest.getSpaceId())
                 .orElseThrow(NoSuchSpaceException::new);
         Reservation reservation = reservations
                 .findById(reservationId)
                 .orElseThrow(NoSuchReservationException::new);
 
-        checkCorrectPassword(reservation, reservationCreateUpdateWithPasswordRequest.getPassword());
-        doDirtyCheck(reservation, reservationCreateUpdateWithPasswordRequest, space);
-        validateAvailability(space, reservationCreateUpdateWithPasswordRequest, reservation);
+        doDirtyCheck(reservation, reservationCreateUpdateRequest, space);
+        validateAvailability(space, reservationCreateUpdateRequest, reservation);
 
-        reservation.update(reservationCreateUpdateWithPasswordRequest, space);
+        reservation.update(reservationCreateUpdateRequest, space);
         reservations.save(reservation);
         return SlackResponse.from(reservation);
     }
@@ -88,20 +92,25 @@ public class UserReservationService extends ReservationService {
     public SlackResponse deleteReservation(
             final Long mapId,
             final Long reservationId,
-            final ReservationPasswordAuthenticationRequest reservationPasswordAuthenticationRequest) {
+            final Member provider) {
         validateMapExistence(mapId);
+        validateAuthorityOnMap(mapId, provider);
 
         Reservation reservation = reservations
                 .findById(reservationId)
                 .orElseThrow(NoSuchReservationException::new);
-        checkCorrectPassword(reservation, reservationPasswordAuthenticationRequest.getPassword());
         reservations.delete(reservation);
         return SlackResponse.from(reservation);
     }
 
-    private void checkCorrectPassword(final Reservation reservation, final String password) {
-        if (reservation.isWrongPassword(password)) {
-            throw new ReservationPasswordException();
+    private void validateAuthorityOnMap(final Long mapId, final Member provider) {
+        List<Map> providerMaps = maps.findAllByMember(provider);
+        if (noMapsMatch(providerMaps, mapId)) {
+            throw new NoAuthorityOnMapException();
         }
+    }
+
+    private boolean noMapsMatch(final List<Map> providerMaps, final Long mapId) {
+        return providerMaps.stream().noneMatch(map -> map.hasSameId(mapId));
     }
 }
