+++ backend/src/test/java/com/woowacourse/zzimkkong/CommonFixture.java
@@ -4,7 +4,6 @@ import com.woowacourse.zzimkkong.domain.Map;
 import com.woowacourse.zzimkkong.domain.Member;
 import com.woowacourse.zzimkkong.domain.Reservation;
 import com.woowacourse.zzimkkong.domain.Space;
-import com.woowacourse.zzimkkong.dto.reservation.CoordinateResponse;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
